Atlas progress dialog adapted to the new download process
git-svn-id: https://mobac.svn.sourceforge.net/svnroot/mobac/trunk@339 83dc79ef-cac3-4f51-90f7-e9e68c987b9c

+++ src/tac/program/interfaces/AtlasInterface.java
@@ -20,5 +20,5 @@ public interface AtlasInterface extends CapabilityRenameable, Iterable<LayerInte
 
 	public AtlasOutputFormat getOutputFormat();
 
-	public long calculateTilesToDownload();
+	public int calculateTilesToDownload();
 }
